requirements.txt
pinecone
openai
python-dotenv
fastapi[standard]
uvicorn
python-multipart
